+++ web/app/controllers.php
@@ -27,15 +27,7 @@ $app->before(function(Request $request, Application $app) {
 
 $controllers = $app['controllers_factory'];
 $controllers->get('/', function () use ($app) {
-    // Load fullcalendar language pack
-    $scripts = $app['scripts'];
-    $locale = $app['locale'];
-
-    if (isset($app['fullcalendar.languages'][$locale]) && !empty($app['fullcalendar.languages'][$locale])) {
-        $scripts[] = 'fullcalendar-locale/' . $app['fullcalendar.languages'][$locale] . '.js';
-    }
-
-    return $app['twig']->render('calendar.html', [ 'scripts' => $scripts ]);
+    return $app['twig']->render('calendar.html');
 })
 ->bind('calendar');