+++ web/app/services.php
@@ -187,6 +187,30 @@ $app['csrf.manager'] = $app->share(function ($app) {
     return new Symfony\Component\Security\Csrf\CsrfTokenManager(null, $storage);
 });
 
+// Asset manager
+$app['assets.packages'] = $app->share(function ($app) {
+    $strategy = $app['assets.strategy'];
+    // Required to work under a non-root path
+    $stack = new Symfony\Component\Asset\Context\RequestStackContext($app['request_stack']);
+
+    $default_package = new Symfony\Component\Asset\Package($strategy);
+
+    $packages = [
+        'css' => new Symfony\Component\Asset\PathPackage('/dist/css', $strategy, $stack),
+        'js' => new Symfony\Component\Asset\PathPackage('/dist/js', $strategy, $stack),
+    ];
+
+    return new Symfony\Component\Asset\Packages($default_package, $packages);
+});
+
+$app['assets.strategy'] = $app->share(function($app) {
+    if ($app['environment'] === 'dev') {
+        return new Symfony\Component\Asset\VersionStrategy\EmptyVersionStrategy();
+    }
+
+    return new Symfony\Component\Asset\VersionStrategy\StaticVersionStrategy('v' . AgenDAV\Version::V);
+});
+
 
 // Sharing support enabled
 if ($app['calendar.sharing'] === true) {